make helper enums public - alpha .1
